requirements.txt
pandas
openpyxl
pyautogui
pywin32

python-dotenv
